Move APIError to separate file
